+++ src/features/viewProblem/webview.ts
@@ -6,14 +6,9 @@ import useWebviewResponseHandle from '@/utils/webviewResponse';
 import { checkCPH, sendCphMessage } from './cph';
 import jumpToCphEventEmitter from './jumpToCphEventEmitter';
 import { tagManager } from '@/utils/tagManager';
-import {
-  getSubmissionContext,
-  selectOpenDocument,
-  submitDocument
-} from '@/features/submit';
+import { createSubmissionControls } from './submissionControls';
 
 export default async function showProblemWebview(data: ProblemData) {
-  let selectedDocument = vscode.window.activeTextEditor?.document;
   const panel = vscode.window.createWebviewPanel(
     'luogu.problemPanel',
     `${data.problem.pid} ${data.problem.title ?? data.problem.content.name}`,
@@ -25,63 +20,18 @@ export default async function showProblemWebview(data: ProblemData) {
       enableCommandUris: ['luogu.solution']
     }
   );
+  const submissionControls = createSubmissionControls(panel, data);
   useWebviewResponseHandle(panel.webview, {
     checkCph: checkCPH,
     jumpToCph: () => sendCphMessage(data),
-    getSubmissionContext: () =>
-      getSubmissionContext(selectedDocument, data.lastLanguage),
-    submitProblem: async ({ language }) => {
-      const problem = {
-        pid: data.problem.pid,
-        cid: data.contest?.id
-      };
-      const submissionContext = getSubmissionContext(
-        selectedDocument,
-        data.lastLanguage
-      );
-      const selectedLanguage = submissionContext.languages.find(
-        option => option.label === language
-      );
-      if (!selectedLanguage)
-        return vscode.commands.executeCommand<boolean>(
-          'luogu.sumbitCode',
-          problem
-        );
-      let submissionDocument = selectedDocument;
-      if (!submissionDocument || submissionDocument.isClosed) {
-        const selectedEditor = await selectOpenDocument();
-        if (!selectedEditor) return false;
-        submissionDocument = selectedEditor.document;
-      }
-      return submitDocument(problem, submissionDocument, selectedLanguage);
-    }
+    ...submissionControls.handlers
   });
-  const postSubmissionContext = () =>
-    panel.webview.postMessage({
-      type: 'submissionContextChanged',
-      data: getSubmissionContext(selectedDocument, data.lastLanguage)
-    });
-  const activeEditorListener = vscode.window.onDidChangeActiveTextEditor(
-    editor => {
-      if (!editor) return;
-      selectedDocument = editor.document;
-      void postSubmissionContext();
-    }
-  );
-  const closeDocumentListener = vscode.workspace.onDidCloseTextDocument(
-    document => {
-      if (document !== selectedDocument) return;
-      selectedDocument = undefined;
-      void postSubmissionContext();
-    }
-  );
   const jumpToCphListener = jumpToCphEventEmitter.event(() => {
     if (panel.active) sendCphMessage(data);
   });
   panel.onDidDispose(() => {
     jumpToCphListener.dispose();
-    activeEditorListener.dispose();
-    closeDocumentListener.dispose();
+    submissionControls.dispose();
   });
   const tagsArray = Array.from((await tagManager.getAllTags()).values());
   panel.webview.html = getReactWebviewHtml(
