+++ webview/views/viewProblem/submissionControls.css
@@ -1,0 +1,4 @@
+.submissionControls {
+  display: flex;
+  gap: 5px;
+}
